fz.easycode_action, fz.easycodetouch_action and tz.easycode_auto_relock
